fix clang-tidy

+++ power_grid_model_c/power_grid_model/include/power_grid_model/auxiliary/serialization/serializer.hpp
@@ -254,7 +254,7 @@ class Serializer {
     void pack_element_in_dict(void const* element_ptr, ComponentBuffer const& component_buffer) {
         uint32_t valid_attributes_count = 0;
         for (auto const& attribute : component_buffer.component->attributes) {
-            valid_attributes_count += !check_nan(element_ptr, attribute);
+            valid_attributes_count += static_cast<uint32_t>(!check_nan(element_ptr, attribute));
         }
         packer_.pack_map(valid_attributes_count);
         for (auto const& attribute : component_buffer.component->attributes) {
